requirements.txt
# OpenGradient ALPHA OG - Persistent AI Agent
# Core SDK and LangChain integration
opengradient>=0.4.7,<0.5.0
langchain-opengradient>=0.1.2
langchain-core>=0.3.15,<0.4.0

# HTTP and env
requests>=2.28.0
python-dotenv>=1.0.0

# CLI colors (Windows-safe)
colorama>=0.4.6

# Dashboard (Streamlit)
streamlit>=1.28.0
plotly>=5.18.0

# Optional: for richer CLI
readline; sys_platform != "win32"
